Fix Tracer and Flow auto-configurations
