--- riptide-spring-boot-autoconfigure/src/main/java/org/zalando/riptide/autoconfigure/TracerAutoConfiguration.java
@@ -1,12 +1,0 @@
-package org.zalando.riptide.autoconfigure;
-
-import io.opentracing.Tracer;
-import io.opentracing.noop.NoopTracerFactory;
-import org.springframework.context.annotation.Bean;
-
-public class TracerAutoConfiguration {
-    @Bean
-    public Tracer tracer() {
-        return NoopTracerFactory.create();
-    }
-}
