+++ riptide-spring-boot-autoconfigure/src/test/java/org/zalando/riptide/autoconfigure/OpenTracingTestAutoConfiguration.java
@@ -5,7 +5,6 @@ import io.opentracing.mock.MockTracer;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.zalando.opentracing.flowid.autoconfigure.OpenTracingFlowIdAutoConfiguration;
 
 @Configuration
 @AutoConfigureBefore(OpenTracingFlowIdAutoConfiguration.class)
